Reorganize project structure and modularize verify feature
