+++ scripts/cleanup-commands.js
@@ -1,0 +1,1 @@
+// Platzhalter
