+++ src/modules/verify/config.js
